# api-spec.yaml
openapi: 3.0.1
info:
  title: CirendsAPI
  version: "1.0"
servers:
  - url: https://cirends.runasp.net/api
  - url: http://localhost:5000/api
...
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
      description: "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
...
paths:
...
  /activities/{activityId}/tasks:
    get:
      tags: [Tasks]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                - { id: 11, name: "Nupirkti maisto", status: 0, priority: 1, activityId: 1 }
        "404": { description: Not Found }
    post:
      tags: [Tasks]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            example: { name: "Nupirkti maisto", description: "…", dueDate: "2025-01-10T12:00:00Z", priority: 1 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              example: { id: 11, name: "Nupirkti maisto", status: 0, priority: 1, activityId: 1 }
        "400": { description: Bad Request }
        "404": { description: Not Found }
  /activities/{activityId}/tasks/{id}:
    get:
      tags: [Tasks]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              example: { id: 11, name: "Nupirkti maisto", status: 1, activityId: 1 }
        "404": { description: Not Found }
    put:
      tags: [Tasks]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            example: { status: 2, name: "Nupirkti maisto" }
      responses:
        "204": { description: No Content }
        "400": { description: Bad Request }
        "404": { description: Not Found }
    delete:
      tags: [Tasks]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
        "400": { description: Bad Request }
        "404": { description: Not Found }
...